Add some New method for logic
