+++ bin/board.rb
@@ -8,7 +8,7 @@ class Board
     while i < 9
       j = 0
       while j < 3
-        print " #{board[i]} "
+        print " #{@board[i]} "
         print '|' unless j == 2
         j += 1
         i += 1
@@ -20,8 +20,8 @@ class Board
   def replace(destination, source)
     i = 0
     while i < 9
-      if board[i] == destination
-        board[i] = source
+      if @board[i] == destination
+        @board[i] = source
         return true
       end
       i += 1
