+++ bin/game_logic.rb
@@ -1,4 +1,7 @@
 module GameLogic
+
+  public 
+  
   def check_valid_move(user_input, symbol, board)
     return false if user_input.negative? && user_input > 9
 
@@ -13,28 +16,7 @@ module GameLogic
     false
   end
 
-  def play_turn(user, symbol, board)
-    wrong_move = false
-    player_turns = false
-    while player_turns == false
-      case wrong_move
-      when false
-        puts "#{user}! it is your turn"
-      when true
-        puts "#{user}! Wrong Move Try again!"
-      end
-      user_input = gets.chomp.to_i
-      check_valid = check_valid_move(user_input, symbol, board)
-      if check_valid
-        player_turns = true
-      else
-        wrong_move = true
-      end
-    end
-    false
-  end
+  def is_winner
 
-  def check_winner()
-    
   end
 end
